fix(): Delete goal if no IK can be found
When a goal was present in map frame and we kill the localization
(world_model). The head ref server was going nuts since it could not
perform an inververse kinematics from the map position to the pan / tilt
joint positions. This will make sure that if thist transformation goes
wrong, the goal will be removed.